+++ packages/pv-stylemark/config/default.config.js
@@ -1,11 +1,9 @@
 const defaultConfig = {
   destPath: "target",
   cdTemplatesSrc: "src/templates/",
-  lsgTemplatesSrc: "src/styleguide/templates/",
   componentsSrc: "src/components/",
   cdPagesSrc: "src/pages/",
   hbsHelperSrc: "helpers/handlebarsHelper/",
-  lsgAssetsSrc: "src/assets/",
   lsgIndex: "src/styleguide/index.html",
   lsgConfigPath: "config/config.stylemark.yaml",
   resourcesSrc: "resources",