+++ packages/pv-stylemark/scripts/dev.js
@@ -1,98 +1,46 @@
-const gulp = require("gulp");
+const { watchFile } = require("fs-extra");
 
+const { watchGlob } = require("../helper/io-helper");
+const { resolveApp, getAppConfig, join } = require("../helper/paths");
 // Assemble Clickdummy
-const {
-  assembleClickdummyComponents,
-} = require("../gulp-tasks/clickdummy_tasks/assembleClickdummyComponents");
-const {
-  assembleClickdummyPages,
-} = require("../gulp-tasks/clickdummy_tasks/assembleClickdummyPages");
-const {
-  copyClickdummyFiles,
-} = require("../gulp-tasks/clickdummy_tasks/copyClickdummyFiles");
+const { assembleClickdummyComponents } = require("../tasks/clickdummy/assembleClickdummyComponents");
+const { assembleClickdummyPages } = require("../tasks/clickdummy/assembleClickdummyPages");
+const { copyClickdummyFiles } = require("../tasks/clickdummy/copyClickdummyFiles");
 // Assemble Stylemark
-const {
-  assembleLSGComponents,
-} = require("../gulp-tasks/lsg_tasks/assembleLSGComponents");
-const {
-  copyStyleguideFiles,
-} = require("../gulp-tasks/lsg_tasks/copyStyleguideFiles");
-const { buildStylemark } = require("../gulp-tasks/lsg_tasks/buildStylemark");
-const { getAppConfig, join } = require("../helper/paths");
+const { buildDDS } = require("../tasks/lsg/buildDDS");
 
-const recompileMessage = (done) => {
-  console.log("Recompiling LSG...");
-  done();
+const buildClickdummy = async () => {
+  await Promise.all([assembleClickdummyComponents(), assembleClickdummyPages(), copyClickdummyFiles()]);
 };
 
-const recompiledMessage = (done) => {
-  console.log("LSG Recompiled!");
-  done();
+const buildLSG = async () => {
+  await buildDDS();
 };
 
-const buildClickdummy = (done) => {
-  return gulp.series(
-    assembleClickdummyComponents,
-    assembleClickdummyPages,
-    copyClickdummyFiles
-  )(done);
+const buildAll = async () => {
+  console.log("Assemble LSG...");
+  await buildClickdummy();
+  await buildLSG();
+  console.log("LSG assembled!");
 };
-
-const buildLSG = (done) => {
-  return gulp.series(
-    assembleLSGComponents,
-    copyStyleguideFiles,
-    buildStylemark
-  )(done);
-};
-
-const { componentsSrc, cdPagesSrc, lsgAssetsSrc, lsgIndex } = getAppConfig();
 
 const watchFiles = () => {
-  gulp.watch(
-    join(componentsSrc, "**/*.md"),
-    gulp.series(
-      recompileMessage,
-      copyStyleguideFiles,
-      buildStylemark,
-      recompiledMessage
-    )
-  );
-  gulp.watch(
-    join(cdPagesSrc, "**/*.hbs"),
-    gulp.series(recompileMessage, assembleClickdummyPages, recompiledMessage)
-  );
-  gulp.watch(
-    join(lsgAssetsSrc, "**/*.*"),
-    gulp.series(recompileMessage, copyStyleguideFiles, recompiledMessage)
-  );
-  gulp.watch(
-    lsgIndex,
-    gulp.series(recompileMessage, copyClickdummyFiles, recompiledMessage)
-  );
-  for (const ext of ["hbs", "json", "yaml", "yml"]) {
-    gulp.watch(
-      join(componentsSrc, `**/*.${ext}`),
-      gulp.series(
-        recompileMessage,
-        assembleClickdummyComponents,
-        assembleClickdummyPages,
-        assembleLSGComponents,
-        buildStylemark,
-        recompiledMessage
-      )
-    );
-  }
-};
+  const { componentsSrc, cdPagesSrc, lsgIndex } = getAppConfig();
+  watchGlob(resolveApp(join(componentsSrc, "**/*.md")), () => buildLSG());
+  watchGlob(resolveApp(join(cdPagesSrc, "**/*.hbs")), () => assembleClickdummyPages());
+  watchFile(resolveApp(lsgIndex), () => copyClickdummyFiles());
 
-const build = (done) => {
-  return gulp.series(buildClickdummy, buildLSG)(done);
+  for (const ext of ["hbs", "json", "yaml", "yml"]) {
+    watchGlob(resolveApp(join(componentsSrc, `**/*.${ext}`)), () => buildAll());
+  }
 };
 
 console.log("Assemble LSG in dev-mode...");
 
-build(() => {
-  console.log("LSG compiled!");
+const startDev = async () => {
+  await buildAll();
   console.log("Start LSG watch...");
   watchFiles();
-});
+};
+
+startDev();