+++ packages/custom-elements-data-extractor/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/pro-vision/fe-tools/compare/@pro-vision/custom-elements-data-extractor@0.2.3...@pro-vision/custom-elements-data-extractor@0.2.4) (2023-04-14)
+
+**Note:** Version bump only for package @pro-vision/custom-elements-data-extractor
+
+
+
+
+
 ## [0.2.3](https://github.com/pro-vision/fe-tools/compare/@pro-vision/custom-elements-data-extractor@0.2.2...@pro-vision/custom-elements-data-extractor@0.2.3) (2022-05-06)
 
 **Note:** Version bump only for package @pro-vision/custom-elements-data-extractor