+++ packages/pv-create-component/CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-create-component@0.1.2...@pro-vision/pv-create-component@0.2.0) (2023-04-14)
+
+
+### Features
+
+* **pv-create-component:** allow pv-create-component to be configured ([395d2c1](https://github.com/pro-vision/fe-tools/commit/395d2c106b92943a8f59ba2acab2894c642b3fea))
+
+
+### BREAKING CHANGES
+
+* **pv-create-component:** the node api has slightly changed, also the "unitType" property from templates
+argument is renamed to "unit"
+
+
+
+
+
+## [0.1.2](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-create-component@0.1.1...@pro-vision/pv-create-component@0.1.2) (2022-05-13)
+
+**Note:** Version bump only for package @pro-vision/pv-create-component
+
+
+
+
+
 ## [0.1.1](https://github.com/pro-vision/fe-tools/compare/@pro-vision/pv-create-component@0.1.0...@pro-vision/pv-create-component@0.1.1) (2022-05-06)
 
 **Note:** Version bump only for package @pro-vision/pv-create-component